Refactor settings account navigation
