+++ apps/web/public/locales/en/settings.json
@@ -390,6 +390,7 @@
     }
   },
   "account": {
+    "title": "Account",
     "actions": {
       "change": "Change"
     },
